requirements.txt
asgiref==3.2.7
astroid==2.4.2
beautifulsoup4==4.9.1
Django==3.0.2
django-background-tasks==1.2.5
django-bootstrap4==1.1.1
django-compat==1.0.15
django-cors-headers==3.2.1
django-crispy-forms==1.8.1
django-filter==2.2.0
django-rest-auth==0.9.5
django-simple-history==2.11.0
django-tables2==2.3.1
djangorestframework==3.11.0
html5lib==1.0.1
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
mysqlclient==1.4.6
Pillow==7.0.0
pylint==2.5.3
pylint-plugin-utils==0.6
PyPDF2==1.26.0
pytz==2020.1
reportlab==3.5.42
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
toml==0.10.1
webencodings==0.5.1
whitenoise==5.0.1
wrapt==1.12.1
xhtml2pdf==0.2.4
